# backend/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Sarvekshanam API
  description: |
    **Sarvekshanam** — Multi-system Security Operations Platform.

    ## Authentication
    All API routes (except `/auth/register` and `/auth/login`) require a `Bearer <token>` header.
    Obtain a token via `POST /auth/login`.

    ## WebSocket
    Connect to `ws://<host>/ws` and send:
    ```json
    { "type": "AUTH", "token": "<jwt>" }
    { "type": "SUBSCRIBE", "sessionId": "<id>" }
    ```
  version: 1.0.0
  contact:
    name: Sarvekshanam
servers:
  - url: /
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Success:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:    { type: object }
...
    User:
      type: object
      properties:
        id:         { type: string, format: uuid }
        username:   { type: string }
        role:       { type: string, enum: [viewer, admin] }
        created_at: { type: string, format: date-time }
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account (viewer role)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, minLength: 3 }
                password: { type: string, minLength: 6 }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token: { type: string }
                          user:  { $ref: '#/components/schemas/User' }
        '400': { description: Validation error }
        '409': { description: Username taken }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token: { type: string }
                          user:  { $ref: '#/components/schemas/User' }
        '401': { description: Bad credentials }
...
  /auth/oidc:
    get:
      tags: [Auth]
      summary: Initiate OIDC SSO flow (redirects to IdP)
      security: []
      responses:
        '302': { description: Redirect to Identity Provider }
        '404': { description: OIDC not configured }
  /auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC SSO callback from IdP
      security: []
      responses:
        '302': { description: Redirect to frontend with JWT token or error }
        '404': { description: OIDC not configured }

  # ── Health ─────────────────────────────────────────────────────────────────
...